회원가입, 로그인, 중복체크 구현 완료

+++ src/main/java/kr/gg/compick/match/dao/MatchRepository.java
@@ -2,8 +2,8 @@ package kr.gg.compick.match.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.fasterxml.jackson.databind.deser.DataFormatReaders.Match;
+import kr.gg.compick.domain.Match;
 
-public interface MatchRepository extends JpaRepository<Match, Integer> {
+public interface MatchRepository extends JpaRepository<Match, Long> {
 
 }
